Add files via upload
This project's fundamental goal is to build an electronic voting machine that will contribute to the eradication of vote fraud in manual systems and earlier iterations of electronic voting. The thesis investigates and suggests a technique with numerous tiers of verifications to guarantee the device's dependability. Each voter is placed into the system only after being recognized and compared with the provided database of registered voters thanks to the use of a biometric fingerprint sensor. The voter will be able to proceed with selecting their favorite candidate from the panel of buttons once the corresponding fingerprint is matched with the data given. After that, an LCD displays the results so that voters can see them. The proposed project displays transparency and also carries the feature of being autonomous during the course of operation.